Added .streamadd <url> which will figure out which platform the url is from and subscribe to it.
Completely reworked how stream notifications work.
Stream notifications will now be only checked on the 0th shard and saved to redis.
When a change occurs (which are now checked for every 20 seconds) it will publish the array of checked statuses to redis.
All subscribers (all shards) will go through the list of updated statuses, get the channels in which that stream is subscribed to, and post the status change message.
The subscribers only know which shards they were subscribed to. As everything is stored in redis, there is no duplication anymore.