Added --clean option for .xplb and .lb to only show users on this server. Keep...
Added --clean option for .xplb and .lb to only show users on this server. Keep in mind it will only use top 10000 users in both cases for filtering. If you or your guild members are below #10000 on the .lb, you won't show up on the .lb --clean for performance reasons. The users will be redownloaded to show the most accurate results up to once per hour.